Avoid prism-import.html, it breaks demo document styles in IE & Edge
There is a `<link rel="stylesheet">` stylesheet include in
“prism-element/prism-import.html”. Because of the polyfill issue, this
breaks document-level styles in IE and Edge, see:
https://github.com/webcomponents/webcomponentsjs/issues/799.

This workaround removes that import from `<nice-demo-snippet>`. Instead
of using `<prism-highlighter>`, `<nice-demo-snippet>` implements
the same and invokes `Prism.highlight` itself.

This allows to get the demo pages look nice in IE and Edge, as long as
`<nice-demo-snippet>` is used for all demos.

Note, that the original `<demo-snippet>` suffers from the same issue.
We should avoid importing it as well to preserve the nice pages look
in IE and Edge.